Created a better entity checker that allows me to use a diff range
also created a sub tag script and put it on the bosses and gave them Boss1 and Boss2 tags respectively.
