+++ Assets/Prefabs/Enemies/Bosses/Priest.prefab
@@ -18,6 +18,7 @@ GameObject:
   - component: {fileID: 8950207567973540252}
   - component: {fileID: 5706574705448098848}
   - component: {fileID: 5589102479190657725}
+  - component: {fileID: -8905556615368017591}
   m_Layer: 7
   m_Name: Priest
   m_TagString: Enemy
@@ -128,6 +129,10 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   maxHealth: 360
+  baseSize: 6
+  numberSizeMultiply: 3
+  minRandomMultiplyAmount: 0.4
+  maxRandomMultiplyAmount: 0.4
   topColor: {r: 1, g: 0.8, b: 0.1, a: 1}
   bottomColor: {r: 1, g: 0, b: 0, a: 1}
 --- !u!114 &-9150303543506523908
@@ -322,6 +327,20 @@ CharacterController:
   m_SkinWidth: 0.08
   m_MinMoveDistance: 0.001
   m_Center: {x: 0, y: 0, z: 0}
+--- !u!114 &-8905556615368017591
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 66097493073879181}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 93a83d44b83947a4985b07b5d764f69b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  subTags:
+  - Boss2
 --- !u!1 &1407740438247208380
 GameObject:
   m_ObjectHideFlags: 0
