+++ Assets/Programming/Character/EntityDetectNearby.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 10a1cfb1c2cfb9b4091d827273a9d698
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
