+++ Assets/Programming/Health System/Health.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c369e918ef61bcb418ac140dab0f87e7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
