+++ Assets/Programming/Health System/HealthWithShield.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9fdbc7e2400dda745ad4d0aa673350c1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
