--- Assets/Programming/HealthSystem.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9e66f86f7edd1d943af4d445cae49faa
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
