--- Assets/Programming/HealthSystem/HealthWithBasicShield.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 75afc6843d2c0e7488b9f3b039df49f6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
