+++ Assets/Programming/Sub Tag System.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 03558f54ad8d9d1468d3cda5b2b4706f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
