+++ Assets/Programming/Sub Tag System/SubTagSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93a83d44b83947a4985b07b5d764f69b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
