+++ Assets/Programming/UI/Player/Minimap/MinimapPointer.cs
@@ -7,23 +7,10 @@ public class MinimapPointer : MonoBehaviour
     [Tooltip("Leave blank to reset")]
     public Transform objectToPointTo;
 
-    private GameObject playerCanvasObject;
-
-    void Start()
-    {
-        playerCanvasObject = PlayerCanvasReference.instance.GetPlayerCanvasReference();
-    }
-
-    // void OnTriggerEnter(Collider other)
-    // {
-    //     if (!other.gameObject.CompareTag("Player")) return;
-
-    //     playerCanvasObject.GetComponent<MinimapController>().targetPoint = objectToPointTo.position;
-    // }
-
-
     public void PointToLocation()
     {
-        playerCanvasObject.GetComponent<MinimapController>().targetPoint = objectToPointTo.position;
+        if (MinimapController.instance == null) return;
+
+        MinimapController.instance.GetComponent<MinimapController>().targetPoint = objectToPointTo.position;
     }
 }
